Remove console.log and move files
