+++ src/containers/BooksList.js
@@ -29,7 +29,6 @@ BooksList.defaultProps = {
 };
 
 function mapStateToProps(state) {
-  console.log(state);
   return {
     books: state.booksReducer,
   };
