+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './styles/index.css';
 import { Provider } from 'react-redux';
-import App from './containers/App';
+import App from './components/App';
 import store from './redux/store';
 
 ReactDOM.render(
